fix: Hero, get_tile_foot, go_to_tile

+++ sprite_objects.py
@@ -26,7 +26,7 @@ class Rectangle(Sprite):
 
 
 class Entity(Sprite):
-    def __init__(self, pos: tuple[int, int], image: Surface, group: Group):
+    def __init__(self, pos: tuple[float, float], image: Surface, group: Group):
         super().__init__(group)
         self.image = image
         self.rect = self.image.get_rect(topleft=pos)
